+++ src/solver/variable/include/antares/solver/variable/economy/STStorageWithdrawalByCluster.h
@@ -63,7 +63,7 @@ struct VCardSTstorageWithdrawalByCluster
     //! Decimal precision
     static constexpr uint8_t decimal =   0;
     //! Number of columns used by the variable
-    static constexpr uint8_t columnCount =   Category::dynamicColumns;
+    static constexpr int columnCount =    Category::dynamicColumns;
     //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate =   Category::spatialAggregateSum;
     static constexpr uint8_t spatialAggregateMode =   Category::spatialAggregateEachYear;
