+++ src/solver/variable/include/antares/solver/variable/economy/dispatchableGeneration.h
@@ -73,7 +73,7 @@ struct VCardDispatchableGeneration
     //! Decimal precision
     static constexpr uint8_t decimal =   0;
     //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr uint8_t columnCount =   10;
+    static constexpr int columnCount =    10;
     //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate =   Category::spatialAggregateSum;
     static constexpr uint8_t spatialAggregateMode =   Category::spatialAggregateEachYear;
