+++ src/solver/variable/include/antares/solver/variable/economy/price.h
@@ -73,7 +73,7 @@ struct VCardPrice
     //! Decimal precision
     static constexpr uint8_t decimal =   2;
     //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr uint8_t columnCount =   1;
+    static constexpr int columnCount =    1;
     //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate =   Category::spatialAggregateAverage;
     static constexpr uint8_t spatialAggregateMode =   Category::spatialAggregateEachYear;
