+++ src/solver/variable/include/antares/solver/variable/economy/thermalAirPollutantEmissions.h
@@ -68,7 +68,7 @@ struct VCardThermalAirPollutantEmissions
     //! Decimal precision
     static constexpr uint8_t decimal =   0;
     //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr uint8_t columnCount =   Antares::Data::Pollutant::POLLUTANT_MAX;
+    static constexpr int columnCount =    Antares::Data::Pollutant::POLLUTANT_MAX;
     //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate =   Category::spatialAggregateSum;
     static constexpr uint8_t spatialAggregateMode =   Category::spatialAggregateEachYear;
